[iOS][Assistant] Adds detent support to Assistant UI
Adds detent support to Assistant UI.

Key changes:
- Adds `AssistantContainerDetent` model and utility creators (Medium, Large).
- Adds dynamic height calculation to `AssistantContainerViewController` to support layout changes without stale state.
- Implements physics-based snapping with momentum projection for natural
  flinging behavior.
- Adds unit tests for detent snapping and limit enforcement.

Demo:
https://drive.google.com/file/d/1AVLj-65t1LXLAGadj1S9IDttfrUbuqeE/view?usp=drive_link&resourcekey=0-GddKP-K0JqzNUeE-jSL1zw

Bug: 469050167
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7596199
Commit-Queue: Ewann Pellé <ewannpv@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1591486}

+++ ios/chrome/browser/assistant/ui/assistant_container_detent.h
@@ -15,12 +15,12 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 @property(nonatomic, readonly, copy) NSString* identifier;
 
 // The resolved value for the detent.
-@property(nonatomic, readonly) CGFloat value;
+@property(nonatomic, readonly) NSInteger value;
 
-// Creates a new detent with a given identifier and a block providing a float
+// Creates a new detent with a given identifier and a block providing an integer
 // value for the detent.
 - (instancetype)initWithIdentifier:(NSString*)identifier
-                     valueResolver:(CGFloat (^)())valueResolver;
+                     valueResolver:(NSInteger (^)())valueResolver;
 
 // Unavailable.
 - (instancetype)init NS_UNAVAILABLE;
