+++ ios/chrome/browser/assistant/ui/assistant_container_detent.mm
@@ -8,11 +8,11 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 #import "base/check.h"
 
 @implementation AssistantContainerDetent {
-  CGFloat (^_valueResolver)();
+  NSInteger (^_valueResolver)();
 }
 
 - (instancetype)initWithIdentifier:(NSString*)identifier
-                     valueResolver:(CGFloat (^)())valueResolver {
+                     valueResolver:(NSInteger (^)())valueResolver {
   self = [super init];
   if (self) {
     CHECK(valueResolver);
@@ -22,7 +22,7 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
   return self;
 }
 
-- (CGFloat)value {
+- (NSInteger)value {
   return _valueResolver();
 }
